Update tools pages, and components
