+++ src/components/ui/EstSummaryItem.tsx
@@ -1,5 +1,10 @@
 import CheckMarkIcon from "../icons/checkMarkIcon";
 
+/**
+ * Displays a label-value pair with a checkmark icon.
+ * @param label The label text.
+ * @param value The value text.
+ */
 interface EstSummaryItemProps {
   label: string;
   value: string;
@@ -11,7 +16,7 @@ const EstSummaryItem: React.FC<EstSummaryItemProps> = ({ label, value }) => {
       <CheckMarkIcon className="flex-none self-start mt-1 fill-white w-4 h-4" />
       <span className="ml-1 capitalize">
         <span className="font-medium">{label}: </span>
-        {value}
+        <span className="font-bold">{value}</span>
       </span>
     </div>
   );
